Rename: 리스트페이지 이름 프로젝트의 통일성을 위해 변경
